chore: move types to shared.ts
